Solves #117
Used ng-if and ng-class directives to selectively hide/show elements
based on a variable defined in the root scope.
The var is root.sidebarEnabled defined in js/on_run.js. The default
value is false. If the user is authenticated, simply set the value to
true using onEnter :
function($rootScope){$rootScope.root.sidebarEnabled=true;} as a
parameter to your route in js/routes.js.
Also did minor changes to the CSS to fix top alignment