Should solve #327
Issue of freezing was caused due to a bug in the implementation of the
histogram binning algorithm. If the data returned by loklak_server was
smaller than the binCount, intervalLength would be set to 0, causing an
infinite loop. Fixed by returning the labels and data as is if they are
smaller than the binCount.