show domain, not name on wall
If this is shown on a big screen, the user must see immediately where to get the source of the feature that they see. If you google 'loklak' you don't get to loklak.net now. Until google finds us better, I believe it would be wise to show the full domain name.